pyproject.toml
[project]
name = "insightx"
version = "0.1.0"
description = "AI-powered customer feedback insight tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.3",
    "jinja2>=3.1.3",
    "google-genai>=1.59.0",
    "requests>=2.31.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

